update dependency injection
