+++ src/Browscap.php
@@ -38,6 +38,7 @@ use BrowscapPHP\Helper\Converter;
 use BrowscapPHP\Helper\Filesystem;
 use BrowscapPHP\Helper\IniLoader;
 use BrowscapPHP\Helper\Quoter;
+use BrowscapPHP\Parser\ParserInterface;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 use WurflCache\Adapter\AdapterInterface;
@@ -81,7 +82,9 @@ class Browscap
      */
     private $cache = null;
 
-    /** @var \Psr\Log\LoggerInterface */
+    /**
+     * @var @var \Psr\Log\LoggerInterface
+     */
     private $logger = null;
 
     /**
@@ -175,7 +178,7 @@ class Browscap
      *
      * @return \BrowscapPHP\Browscap
      */
-    public function setParser(Parser\ParserInterface $parser)
+    public function setParser(ParserInterface $parser)
     {
         $this->parser = $parser;
 
@@ -185,35 +188,16 @@ class Browscap
     /**
      * returns an instance of the used parser class
      *
-     * @return Parser\ParserInterface
+     * @return \BrowscapPHP\Parser\ParserInterface
      */
     public function getParser()
     {
         if (null === $this->parser) {
-            $this->setParser(new Parser\Ini());
+            $patternHelper = new Parser\Helper\GetPattern($this->getCache(), $this->getLogger());
+            $dataHelper    = new Parser\Helper\GetData($this->getCache(), $this->getLogger(), new Quoter());
+
+            $this->parser = new Parser\Ini($patternHelper, $dataHelper, $this->getFormatter());
         }
-
-        $patternHelper = new Parser\Helper\GetPattern();
-        $dataHelper    = new Parser\Helper\GetData();
-
-        $patternHelper
-            ->setCache($this->getCache())
-            ->setLogger($this->getLogger())
-        ;
-
-        $dataHelper
-            ->setCache($this->getCache())
-            ->setLogger($this->getLogger())
-            ->setQuoter(new Quoter())
-        ;
-
-        $this->parser
-            ->setPatternHelper($patternHelper)
-            ->setDataHelper($dataHelper)
-            ->setFormatter($this->getFormatter())
-            ->setCache($this->getCache())
-            ->setLogger($this->getLogger())
-        ;
 
         return $this->parser;
     }
@@ -328,11 +312,7 @@ class Browscap
             throw new Exception('an error occured while setting the local file', 0, $e);
         }
 
-        $converter = new Converter();
-        $converter
-            ->setLogger($this->getLogger())
-            ->setCache($this->getCache())
-        ;
+        $converter = new Converter($this->getLogger(), $this->getCache());
 
         try {
             $converter->convertString($loader->load());
@@ -348,13 +328,8 @@ class Browscap
      */
     public function convertString($iniString)
     {
-        $converter = new Converter();
-
-        $converter
-            ->setLogger($this->getLogger())
-            ->setCache($this->getCache())
-            ->convertString($iniString)
-        ;
+        $converter = new Converter($this->getLogger(), $this->getCache());
+        $converter->convertString($iniString);
     }
 
     /**
@@ -407,12 +382,7 @@ class Browscap
     {
         $this->getLogger()->debug('started fetching remote file');
 
-        $converter = new Converter();
-
-        $converter
-            ->setLogger($this->getLogger())
-            ->setCache($this->getCache())
-        ;
+        $converter = new Converter($this->getLogger(), $this->getCache());
 
         if (class_exists('\Browscap\Browscap')) {
             $resourceFolder = 'vendor/browscap/browscap/resources/';
